Implement event-driven approach
Previously, the parser itself allocated resources for
HTTP message headers or body. This approach made the parser
less versatile and burdened it with resource allocation
responsibility.

To solve these problems, callbacks were introduced. Now, when an
event occurs, the parser is no longer responsible for
implementing the logic to handle it.