Improve error handling
Introduced the `parser_had_error()` function, which returns `true` if an
error occurred and `false` otherwise.

When an error happens, the parser's `error` field is set with a
numerical code. To make error displaying easier, the `parser_get_error()`
function was also introduced; it returns a string representation of
the error.